Move split to utilities file
